refactor: 💡 getCurrentDateTime 抽至共用util
